requirements.txt
# requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
qdrant-client>=1.7.0
openai>=1.12.0
openai-agents>=0.0.7
google-generativeai>=0.7.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
httpx>=0.26.0
sse-starlette>=1.8.0
hypothesis>=6.92.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
nest-asyncio>=1.6.0
